fixed "warning: format '%lu' expects" message
Under 3.10 686 a format warning is generated.
However on a 3.9 amd64 system no such error is generated.

make -C /lib/modules/3.10-2-686-pae/build M=/home/jeri/ldd/fifo_sysfs modules
make[1]: Entering directory `/usr/src/linux-headers-3.10-2-686-pae'
  CC [M]  /home/jeri/ldd/fifo_sysfs/fifo.o
/home/jeri/ldd/fifo_sysfs/fifo.c: In function ‘read_offset_show’:
/home/jeri/ldd/fifo_sysfs/fifo.c:121:2: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 3 has type ‘int’ [-Wformat]
/home/jeri/ldd/fifo_sysfs/fifo.c: In function ‘write_offset_show’:
/home/jeri/ldd/fifo_sysfs/fifo.c:139:2: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 3 has type ‘int’ [-Wformat]
  Building modules, stage 2.
  MODPOST 1 modules
  CC      /home/jeri/ldd/fifo_sysfs/fifo.mod.o
  LD [M]  /home/jeri/ldd/fifo_sysfs/fifo.ko
make[1]: Leaving directory `/usr/src/linux-headers-3.10-2-686-pae'
jeri@ripley:~/ldd/fifo_sysfs$

The code involves subtracting two pointers.  Since pointers may be different
sizes on different systems the warning message occur on systems
with a different size.  To fix the problem the result is cast
to a size that printf expects.

+++ fifo_sysfs/fifo.c
@@ -118,7 +118,8 @@ static ssize_t read_offset_show(struct device *dev,
 								char *buf)
 {
 	struct fifo_dev *fifo_devp = dev_get_drvdata(dev);
-	return sprintf(buf, "%lu\n", fifo_devp->read_ptr - fifo_devp->fifo_start);
+	return sprintf(buf, "%u\n", (unsigned int) (fifo_devp->read_ptr
+													- fifo_devp->fifo_start));
 }
 
 static ssize_t read_offset_store(struct device *dev,
@@ -136,7 +137,8 @@ static ssize_t write_offset_show(struct device *dev,
 								char *buf)
 {
 	struct fifo_dev *fifo_devp = dev_get_drvdata(dev);
-	return sprintf(buf, "%lu\n", fifo_devp->write_ptr - fifo_devp->fifo_start);
+	return sprintf(buf, "%u\n", (unsigned int) (fifo_devp->write_ptr
+													- fifo_devp->fifo_start));
 }
 
 static ssize_t write_offset_store(struct device *dev,
